+++ include/kernel/infrastructure/i686/firmware/acpi.h
@@ -29,8 +29,8 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef JINUE_KERNEL_INFRASTRUCTURE_I686_DRIVERS_ACPI_H
-#define JINUE_KERNEL_INFRASTRUCTURE_I686_DRIVERS_ACPI_H
+#ifndef JINUE_KERNEL_INFRASTRUCTURE_I686_FIRMWARE_ACPI_H
+#define JINUE_KERNEL_INFRASTRUCTURE_I686_FIRMWARE_ACPI_H
 
 #include <kernel/infrastructure/i686/types.h>
 #include <stdint.h>
