+++ include/kernel/infrastructure/i686/firmware/asm/bios.h
@@ -29,8 +29,8 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef JINUE_KERNEL_INFRASTRUCTURE_I686_DRIVERS_ASM_BIOS_H
-#define JINUE_KERNEL_INFRASTRUCTURE_I686_DRIVERS_ASM_BIOS_H
+#ifndef JINUE_KERNEL_INFRASTRUCTURE_I686_FIRMWARE_ASM_BIOS_H
+#define JINUE_KERNEL_INFRASTRUCTURE_I686_FIRMWARE_ASM_BIOS_H
 
 #define BIOS_BDA_EBDA_SEGMENT   0x40e
