+++ kernel/infrastructure/i686/firmware/acpi.c
@@ -32,8 +32,8 @@
 #include <kernel/infrastructure/acpi/acpi.h>
 #include <kernel/infrastructure/acpi/tables.h>
 #include <kernel/infrastructure/acpi/types.h>
-#include <kernel/infrastructure/i686/drivers/asm/bios.h>
-#include <kernel/infrastructure/i686/drivers/acpi.h>
+#include <kernel/infrastructure/i686/firmware/asm/bios.h>
+#include <kernel/infrastructure/i686/firmware/acpi.h>
 #include <kernel/infrastructure/i686/types.h>
 #include <stdbool.h>
 #include <stdint.h>
